// a3front/src/modules/negotiations/products/configuration/content/generic/serviceDetails/interfaces/service-details-request.interface.ts
export interface ServiceDetailsSchedule {
  day: string;
  start: string | null;
  end: string | null;
  applyAllDay: boolean;
  singleTime: boolean;
  active: boolean;
}

export interface ServiceDetailsRequest {
  id: string | null;
  groupingKeys: {
    operatingLocationKey: string;
    supplierCategoryCode: string;
  };
  content: {
    basicInfo: {
      name: string;
      subTypeCode: string;
      profileCode: string;
    };
    logistics: {
      startPointCodes: string[];
      endPointCodes: string[];
      duration: string;
    };
    operability: {
      mode: string;
      applyAllDay: boolean;
      singleTime: boolean;
      schedules: ServiceDetailsSchedule[];
    };
    status: {
      state: string;
      reason: string;
      clientVisible: boolean;
    };
  };
  completionStatus: string;
}